// lib/typings/types.ts
export enum BuildStatus {
  passed = "passed",
  broken = "broken",
  failed = "failed",
  started = "started",
  canceled = "canceled"
}
// ...
export enum XrayIssueSeverity {
  Major = "Major",
  Minor = "Minor",
  Unknown = "Unknown"
}

export enum XrayIssueType {
  Security = "Security"
}

export enum XrayIssueProvider {
  Snyk = "Snyk",
  JFrog = "JFrog",
  Atomist = "Atomist"
}

export enum XrayPackageType {
  Maven = "Maven",
  Build = "Build"
}
// ...
  export type Variables = {
    sha: string;
    buildId: string;
  };

  export type Query = {
    __typename?: "Query";
    Commit?: Commit[] | null;
  };

  export type Commit = {
    __typename?: "Commit";
    builds?: Builds[] | null;
  };

  export type Builds = {
    __typename?: "Build";
    buildId?: string | null;
    repo?: Repo | null;
    data?: string | null;
    push?: Push | null;
    status?: BuildStatus | null;
  };

  export type Repo = {
    __typename?: "Repo";
    owner?: string | null;
    name?: string | null;
    channels?: Channels[] | null;
  };

  export type Channels = {
    __typename?: "ChatChannel";
    channelId?: string | null;
  };

  export type Push = {
    __typename?: "Push";
    branch?: string | null;
  };
// ...
export namespace FindBuildsForCommit {
  export type Variables = {
    sha: string;
    branch: string;
  };

  export type Query = {
    __typename?: "Query";
    Commit?: Commit[] | null;
  };

  export type Commit = {
    __typename?: "Commit";
    builds?: Builds[] | null;
  };

  export type Builds = {
    __typename?: "Build";
    buildId?: string | null;
    repo?: Repo | null;
    data?: string | null;
    push?: Push | null;
    status?: BuildStatus | null;
  };

  export type Repo = {
    __typename?: "Repo";
    owner?: string | null;
    name?: string | null;
    channels?: Channels[] | null;
  };

  export type Channels = {
    __typename?: "ChatChannel";
    channelId?: string | null;
  };

  export type Push = {
    __typename?: "Push";
    branch?: string | null;
  };
}
export namespace XrayViolations {
  export type Variables = {};

  export type Subscription = {
    __typename?: "Subscription";
    XrayViolation?: XrayViolation[] | null;
  };

  export type XrayViolation = {
    __typename?: "XrayViolation";
    watch_name?: string | null;
    top_severity?: XrayIssueSeverity | null;
    created?: string | null;
    id?: string | null;
    issues?: Issues[] | null;
  };

  export type Issues = {
    __typename?: "XrayIssue";
    severity?: XrayIssueSeverity | null;
    type?: XrayIssueType | null;
    cve?: string | null;
    provider?: XrayIssueProvider | null;
    created?: string | null;
    summary?: string | null;
    description?: string | null;
    impacted_artifacts?: ImpactedArtifacts[] | null;
  };

  export type ImpactedArtifacts = {
    __typename?: "XrayArtifact";
    pkg_type?: XrayPackageType | null;
    path?: string | null;
    sha256?: string | null;
    name?: string | null;
    sha1?: string | null;
    parent_sha?: string | null;
    depth?: number | null;
    display_name?: string | null;
    infected_files?: InfectedFiles[] | null;
  };

  export type InfectedFiles = {
    __typename?: "XrayInfectedFile";
    name?: string | null;
    path?: string | null;
    sha256?: string | null;
    depth?: number | null;
    parent_sha?: string | null;
    display_name?: string | null;
  };
}